+++ src/components/app-header/app-header.jsx
@@ -1,6 +1,6 @@
 import styles from "./app-header.module.css";
 import {Logo, BurgerIcon, ListIcon, ProfileIcon} from '@ya.praktikum/react-developer-burger-ui-components'
-import {NavigationLink} from "../navigation-link/navigation-link";
+import {NavigationLink} from "./navigation-link/navigation-link";
 
 
 const classes = {
