+++ src/components/burger-ingredients/burger-ingredients.jsx
@@ -2,11 +2,11 @@ import React from "react";
 import styles from "./burger-ingredients.module.css";
 import stylesConstr from "../burger-constructor/burger-constructor.module.css";
 import {Tab} from '@ya.praktikum/react-developer-burger-ui-components'
-import {ListCards} from "../list-cards/list-cards";
+import {ListCards} from "./list-cards/list-cards";
 import {optionalArray, optionalFunc, optionalObject} from "../../utils/prop-types";
 
 
-function BurgerIngredients({data, setSelectedIngredients, selectedIngredients, defaultBun}) {
+function BurgerIngredients({data, setSelectedIngredients, selectedIngredients}) {
 
   BurgerIngredients.propTypes = {
     data: optionalArray,
