+++ src/components/burger-ingredients/list-cards/list-cards.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import styles from "./list-cards.module.css";
 import {CardOther} from "../card/card-other";
 import {CardBuns} from "../card/card-buns";
-import {optionalArray, optionalFunc, optionalObject} from "../../utils/prop-types";
+import {optionalArray, optionalFunc, optionalObject} from "../../../utils/prop-types";
 
 function ListCards({data, setSelectedIngredients, selectedIngredients}) {
